Moved input fields to inspector, added icons and updated styling.
